pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcribe-cli"
version = "0.1.0"
description = "CLI tool for transcribing audio and video files using OpenAI Whisper API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Audio Transcription Team" }
]
keywords = ["transcription", "whisper", "audio", "video", "cli", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "openai>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "srt>=3.5.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pip-audit>=2.0.0",
    "pre-commit>=3.0.0",
]
diarization = [
    "pyannote-audio>=3.0",
]

[project.scripts]
transcribe = "transcribe_cli.cli.main:app"

[project.urls]
Homepage = "https://github.com/jmagly/transcribe-cli"
Documentation = "https://github.com/jmagly/transcribe-cli#readme"
Repository = "https://github.com/jmagly/transcribe-cli"
Issues = "https://github.com/jmagly/transcribe-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/transcribe_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/transcribe_cli --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/transcribe_cli"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 60

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv", "venv"]
